fix memory alignment in application_information_t

+++ firmware/application/ui/ui_btngrid.cpp
@@ -138,7 +138,7 @@ void BtnGridView::add_item(const GridItem& new_item, bool inhibit_update) {
     }
 }
 
-void BtnGridView::insert_item(const GridItem& new_item, uint8_t position, bool inhibit_update) {
+void BtnGridView::insert_item(const GridItem& new_item, size_t position, bool inhibit_update) {
     if (!blacklisted_app(new_item)) {
         if (position < menu_items.size()) {
             auto pos_iter = menu_items.begin() + position;
